refactor:TodoPage삭제,handleSubmit 함수 수정
