+++ src/TodoForm.jsx
@@ -3,14 +3,9 @@ function TodoForm({
   todoText,
   handleChange,
 }) {
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    onSubmit(todoText);
-  };
-
   return (
     <div>
-      <form onSubmit={handleSubmit}>
+      <form onSubmit={onSubmit}>
         <input
           type="text"
           placeholder="할 일을 입력해주세요."
